git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@98 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/branches/howard/pre-1.2.9/src/mydns/axfr.c
@@ -322,7 +322,7 @@ axfr(TASK *t) {
   /* Report result */
   gettimeofday(&finish, NULL);
   Debug(_("AXFR: %u records, %u octets, %.3fs"), 
-	total_records, total_octets,
+	(unsigned int)total_records, (unsigned int)total_octets,
 	((finish.tv_sec + finish.tv_usec / 1000000.0) - (start.tv_sec + start.tv_usec / 1000000.0)));
 #endif
   t->qdcount = 1;
